// npm/src/deps/jsr.io/@nshiab/journalism-finance/1.5.24/src/finance/helpers/rentVsBuy/types/mortgagePayment.ts
export type MortgagePayment = {
  paymentId: number;
  payment: number;
  interest: number;
  capital: number;
  balance: number;
  amountPaid: number;
  interestPaid: number;
  capitalPaid: number;
  effectiveInterestRate: number;
  postedInterestRate: number;
  fixedRateAdjustment: number;
  variableRateAdjustment: number;
};